# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {},
        "license": {}
    },
    "paths": {
        "/v1/spider": {
            "post": {
                "description": "insert episode",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Spider"
                ],
                "summary": "Insert episode from URL of MGTV",
                "operationId": "insert episode",
                "parameters": [
                    {
                        "description": "Name",
                        "name": "name",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/app.SpiderRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.SpiderResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/app.SpiderResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "app.SpiderRequest": {
            "type": "object",
            "required": [
                "channel_id",
                "url"
            ],
            "properties": {
                "channel_id": {
                    "type": "integer"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "app.SpiderResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Mgtv"
                    }
                },
                "error": {
                    "type": "object"
                },
                "length": {
                    "type": "integer"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
        "model.Mgtv": {
            "type": "object",
            "properties": {
                "channel_id": {
                    "type": "integer"
                },
                "content_type": {
                    "type": "string"
                },
                "drama_id": {
                    "type": "integer"
                },
                "drama_title": {
                    "type": "string"
                },
                "duration": {
                    "type": "integer"
                },
                "episode_id": {
                    "type": "integer"
                },
                "episode_url": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "is_intact": {
                    "type": "string"
                },
                "is_new": {
                    "type": "string"
                },
                "is_vip": {
                    "type": "string"
                },
                "next_id": {
                    "type": "string"
                },
                "play_counter": {
                    "type": "integer"
                },
                "src_clip_id": {
                    "type": "string"
                },
                "title1": {
                    "type": "string"
                },
                "title2": {
                    "type": "string"
                },
                "title3": {
                    "type": "string"
                },
                "title4": {
                    "type": "string"
                },
                "ts": {
                    "type": "string"
                }
            }
        }
    }
}